requirements.txt
alembic==1.12.0
Babel==2.13.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
speaklater==1.3
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
WTForms==3.1.0
